flask-basic-blog/requirements.txt
click==8.1.3
Flask==2.2.2
Flask-Alembic==2.0.1
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyYAML==6.0
SQLAlchemy==1.4.40
Werkzeug==2.2.2
WTForms==3.0.1
